vsNoteBook_LM/requirements.txt
google-generativeai
python-dotenv
PyMuPDF
